Coinduction: Basic immediate coinduction support
The kernel is extended to support immediate coinduction by the addition
of a new rule, associated to the also new coindClerk' and related helper
predicates. These perform the dual process of immediate induction,
partly by reuse of existing functions, partly by construction of
predicates by analogy. Some of this points must be doublechecked; all
must be refactored. There is much that can be done to reduce both
treatments to a slightly more general, common case, especially as
pertains to the construction of invariants. Still, changes to the kernel
remain very limited, as should be the case.

The FPC now implements the new clerk, coindClerk', and related helpers,
also by analogy. An important limitation is that naming structures
cannot be convincingly derived without access to the naming structures
contained in Gamma, which is one of the bases of the immediate
coinvariants. Because the kernel stores this indexing information,
however, it is not available to us through the certificates, and for
now a blind search relying on a constant bucket "X" is used as a
placeholder. However, it is very important to decide how to best
overcome this limitation.

Simulation examples (reflexivity and transitivity) can be proved with
immediate coinduction and without reference to the definition of the
'step' predicate.