Coinduction: Modularized examples on sim
Both cases of reasoning on simulation are split and cleaned up to cover
both generic reasoning, which does not appeal to the definition of step,
and computational reasoning, which relies on it.

My conjecture is that it suffices to define generic least fixed points
through the false connective, in the following sense: if they occur on
the right, failure is the only option due to the absence of an inference
rule for it; if on the left, the falsehood will be consumed in the
asynchronous phase and, in effect, it will be as if step "had not
existed," so any reasoning relying on its definition must appeal to the
freeze rule, and to it alone.