Minimalist undefined predicates done right
Dealing with undefined predicates the right way is a bit tricky, but in
the end no mystery. The first thing to notice is that constant formulas
do not do what we want them to: either on the left or on the right, both
true and false shortcircuit some of the cases that we rely on to express
what we want. Consequently, these definitions are both incorrect and can
lead us to express the wrong thing.

On the other hand, we could think about trying to hack our way around
the logic and/or the kernel, providing some special syntax or special
cases, constants, etc. None of these are too clean and disturb the
notion that undefined fixed points should look and behave like fixed
points anyway, and the "users" of those fixed points should not really
care about their concrete definitions.

The ultimately minimalist workaround is as follows: we define a new
vacuous constant of the same type as the predicates of our fixed
points (in our case, under the single "one-argument" case) and use this
constant as the body of our undefined fixed points. Put into practice,
this means that unfold predicates may succeed normally, but the
application of the arguments to the body by the kernel can never succeed
and the troublemaker is thus inhibited from occurring.

Current traces look a bit funny because these inhibited unfolds print
out their whole debugging information, as if they succeeded, but the
kernel throws them back immediately afterwards, without need of a
recursive call, so immediately after this abortive unfold on level [N]
another freeze on level [N] must follow, and succeed.