CCS: New implementation and first example
The implementation of CCS in the logic is based on its treatment in the
Abella repository
  http://abella-prover.org/examples/process-calculi/ccs.html
based in turn on Alwen Tiu's thesis.

The tricky bit concerns representation of fixed points in the process
language using our monotype system. As it turns out, the problem is not
as difficult as it may have seemed.