CCS: Some examples, with FPC and predicate tweaks
The battery of examples advances towards completion. The most
significant change concerns the implementation of unfoldRClerk in the
FPC, which was needed for a greatest fixed point for the first time.
For now, it shares the asynchronous counter previously used in exclusive
by unfoldLClerk, and this sharing can give the proof more options in
certain (for now unusual) circumstances. A split can be very much
considered, so that two parameters, one for each type of unfolding,
would result.

Some deficiencies in the definition of the CCS evaluation step have been
corrected as well, which were not detected by the trivial examples of
the previous commit.

In addition, one significant limitation is identified during the proof
of sim_mu_par. Here, we should recall that, by convention, synchronous
implications are being treated right, then left, which works very well
for least fixed points. Nonetheless, it would seem that when a lemma on
a greatest fixed point is involved, as in the present case, the order
must be reversed for the instantations to take place in the correct
order. Of course, we could offer both choices, which is something that
would multiply the degrees of freedom of existing proofs (thus rather
inconvenient, although one can hope that failure would occur soon if
bounds are tight). Alternatively, some form of guidance, enabling and
disabling one or the other, could be devised, at the cost of a more
complex FPC architecture (which, I am tempted to opine, would be
generally speaking of little use).