Properties of bisimulation
In following the properties of transition systems as described in the
Abella website, we make a detour to cover bisimulation in a generic
way.

The main points of interest concern the differences between proof
certificates involving only least fixed points and the modifications
that are imposed by greatest fixed points.