Prog: Foundations of context subsumption [sic?]
To investigate some interesting properties related to contexts and their
manipulation, we look into using our proof checker as the meta-logic
in which to implement an object-logic, thus enabling reasoning about
other proof systems.

Using our monotype representation, we will design an object logic based
around the concepts of contexts and programs. Contexts will contain
atoms of certain types, which for simplicity will be unfolded into the
definition of context explicitly: that is, we will have some "atom
constructors" that will be allowed as members of a context.

On the other hand, we will have a program predicate that will define
derivations of a disjunct set of atoms recursively: in essence, a Horn
clause database (or related logical language) defining for those
recursively defined atoms rules as formulas in our object logic.

Finally, a sequent predicate implements the interpreter for the object
logic treating each connective of the object logic and provides two
base cases: one for atoms which match those in the context and another,
recursive, for atoms that can be found in the program database.

Critically, there can be no overlap between the atoms in both sets for
the system to be consistent.

As indicated above, everything is implemented on top of our monotype i,
which means that some of the type checks in the whiteboard presentation
are abbreviated or elided for compactness (given that they are not
strictly required by the typing system of the meta-logic).

Thus armed, we are prepared to prove some properties about simple
recursive predicates, like plus: namely, the context independence of
such a predicate, using for support a lemma negating the presence of such
(program-defined) predicates in the context. We present examples of
case analysis of this style of proof, as well as a sequentialized
certificate that exploits step unification across all predicted paths
of the derivation.

Finally, the previously undefined falsehood rules must be (trivially)
defined in our FPCs, since we are for the first time trying to prove
negation properties.