Prog: Extension to more general progs
The initial encoding does not easily extend to complex predicates that
use recursive patterns beyond the elementary case of, say, sum: take,
for example, multiplication, whose recursive case involves a conjunction
of two atoms (plus and mult), themselves possibly recursive.

This pattern is troublesome in proofs by induction which expect the
result of recursion to follow the same pattern as the original
arguments. Therefore, if we try to prove an equivalent property of
context independence for mult, the recursive case matches a complex goal
against the pattern of the original call that is being induced upon (a
single instance of mult), which means that our induction is too weak for
this case.

We get around this difficulty substituting the initial object logic by
an alternative representation of Horn-like programs based on sets: thus,
a program associates a single atom with a set of sub-atoms, and a
sequent takes two lists of atoms: the context and the Horn-like goal.
Here, an empty set is vacuously true, and an arbitrary set is processed
recursively, each atom being assigned in turn to one of the two
recursive cases: either a context atom which succeeds and continues the
derivation on the rest of the goal set, or a program atom that generates
a list of sub-goals that replace it in the list before proceeding
recursively.

Because we are allowed to mix modes of deduction and everything is
represented through uniform sets, derivation patterns are simplified (it
was dubious that we would need disjunction or implication in the object
logic, at least at this stage, anyway).

The model thus amended proves the previous properties without
modification. Furthermore, the equivalent properties for multiplication
are verified by exactly the same certificates (using the adequate
context lemma).