Certificates on plus adjusted to new limits format
On the one hand, this is something that should be done immediately
upon branch merges, as it creates a bunch of trivial regressions.
On the other, given the nature of the changes, it should not be so
tedious, and a more "declarative" format that assumes zero defaults
and declares allowances explicitly would go a long way toward
making this easier, as well as making certificates easier to read.