Undo optimization on freezeL certificate types
Previously, we had decided to restrict freeze to the case where no
pending inductions were present in order to enforce a certain,
rather strict, discipline of induction "first and only," avoiding
thus a certain kind of ambiguity and the maligned choice that
comes with it.

Under certain assumptions this is no longer valid, and indeed this
will often be the case when lemmas or nested inductions are present
and anything but absolutely trivial manipulations will get us to
the point where we can apply the lemma (hypotheses and conclusion
stored on their respective sides).

Now freezing and induction are in direct competition, a scenario
that had been previously avoided. It may be a good idea to
prioritize freezes, as they are simpler operations and will lead to
failures faster than potentially involved sub-theorems.

To some extent, the introduction of delays may offer some of the
effects of this change, but would presumably make proofs and
formulations significantly more delicate. I would not call this a
good alternative.