Nested: Example on explicit nested induction
Nested induction works out of the box if all nested invariants are
supplied (the top invariant can be left to the system to generate,
given that it has a clean context to work on). Most of the simple
proofs we are considering work with trivial invariants, and their
inclusion in a certificate serves only as a workaround to avoid the
tainted context (Sigma) that is maintained for the sole purpose of
deriving these invariants.

Even assuming for the moment that the context cannot be relied upon,
this necessity places an additional demand on the user to understand
the structure of the proof at the time of application of a lemma.
This should be easy enough if the information being present is just
what is needed for its application, but could get messy if not. A
better approach would be to query the user for an invariant when the
time to apply the lemma is reached.

One relative downside is that manual induction does not assume the
invariant to be correct along the trivial branch, which means that
computational bounds must take this fact into account as well.
Currently, these demands accummulate on the bounds of the base
certificate that enables the manual induction, which consequently
operates in two phases: before the induction and throughout the
trivial branch. This is tricky and could (should?) be split.

A related issue is that Abella proofs should be easy to translate
into tightly bounded certificates, with the main difference
concerning how inductive hypotheses are handled by both systems.
Of particular interest is the fact that Abella will never treat
the trivial branch (and will arrive at the induction step
somewhat differently), so it will be of no assistance to guide this
part of the computation.

It has been suggested that computation of immediate invariants is
not very declarative in nature and it may be wise not to trust it,
i.e., to subject it too to the trivial branch that it is supposed
to fulfill by definition. As regards automatic translation of
proofs, the same dilemma noted above would arise.

Finally, nested immediate inductions do not seem to do anything too
aberrant in spite of the current problems with the context. They
seem, however, to exhibit some kind of looping behavior that needs
to be investigated.