Bedwyr extensions used throughout
Signature files are given the .sig extension, module files the .mod
extension (but a .mod file does not require a .sig unless it wishes or
needs to expose something before its use).

Further refinements are possible, in particular regarding the
definition of FPC files and harness files. However, it is all rather
delicate because Bedwyr's module system is itself weak.