Dummy FPC and examples frankensteined into working
These old fossils are finally brought up to date, with some provisos.
The FPC in particular is given trivial extensions to the simple
propagation rules that were already present in it. Indexes are also
hacked into submission, but reliance on these could lead to troublesome
results. I would not rely on them presently.

Nonetheless, the trivial set of "unit tests" works as it should. Bedwyr
is bad at this sort of thing, so growing this collection into something
more serious seems sadly like a lot of work with very little return of
investment. Still, I'm not happy about relying on "integration tests"
alone.