fpc-simple: continuation certificates with examples
Added to the full, atomic certificates are a number of chained
certificates that add continuation and finer guidance to proof
search, including specification of decisions on local or global
formulas, and specification of computational bounds, as well as the
necessary branching structure to represent decision trees.

Previously resilient examples can be dealt with in a moderate amount
of time, but ambiguities in the choice of hypotheses, especially
those involving reversible properties like commutativity, still have
a large impact on performance, an impact that was previously
tackled adding hypothesis naming in the admin template.

The new certificate constructors restrict the number of parameters
of the monolithic originals and refine those aspects that were
specific to each bipole, and thus not directly approachable by the
block certificates. In particular:

 - 'induction?' applies immediate induction opportunistically and
   without any claim to sophistication, e.g. unfoldings or bipoles.

 - 'apply?' is the main bipole certificate, charged with making a
   decision on a specified index, which can point to either a local
   stored formula or a global lemma name, and has a certain number
   of allowed unfoldings to carry out its task, ending its task
   when the focus is released.

 - 'case?' implements the branching structure, currently only on the
   asynchronous phase (case analysis of clauses implementing a
   predicate in the typical use case), possibly with asynchronous
   unfoldings, and giving continuations for each branch.

Of course, block certificates could be seen as rings in a chain of
certificates as well.

An interesting thing to note is that, if correctly implementic, the
arithmetic to go from chained certificates to block certificates is
clear and straightforward.