elaboration: blocks to chains in simple FPC (pt.I)
First tentative implementation of certificate elaborations. Starting
from a block certificate, return a chained stepwise elaboration with
tight bounds and concrete indexes. To do this, a new constructor
pair# is defined that expects a defined (block) certificate and
returns a concrete elaboration as its second member. To keep track
of bounds and indexes, additional members are added to the structure
for bookkeeping purposes.

It must be borne in mind that pairs, unless the constructor hierarchy
is ranked clearly, could be essentially a free-for-all from any
certificate to its more elaborated form, and could thus be tedious to
maintain; canonicity becomes an very current issue.

For now, we attempt to treat the
  induction -> apply -> start block
by converting it into a tree
  induction? -> ( case? * -> apply? ) * -> start
with some restrictions. Shortcircuits before reaching maximum depth
are currently not handled with grace, and trivial cases (e.g. those
that do not result in different lemma subtrees) are not simplified.

Trivial examples will work under this scheme, but the presence of
numerous logic variables in the certificate pair in course of
elaboration seems to cause (left) unification problems, chiefly
after eqClerk's success and the hypothetical step in the kernel.
However, trees and their splits must be decided dynamically, so there
is no escaping the need to elaborate "on the fly" as opposed to e.g.
leaving linear sequences of lemmas to be determined, something we
practiced with the admin FPC template.

For now, the challenge is to circumscribe this to the available logic
and find any potential weak spots, possibly attempting to minimize
the number of free variables.