elaboration: blocks with decision trees (pt.III)
Deviating slightly from the pure binary decision tree approach, a
natural procedure is identified (if it remains somewhat simplified):
branching events, namely orClerk, introduce branching points, whereas
decisions of any kind introduce a linear extension that details the
type and index chosen and provides a continuation for further work.
No guesses need to be made at the outset, and everything will be
defined at runtime. (Currently, branches are "closed" simply by
leaving them open, and as little specialized code is added as
possible).

Following this discipline, one can go from monolithic certificates to
a pair that returns, in lockstep, this pseudo-binary tree of all
relevant decisions taken. It is not necessarily pretty, nor does it
address the fact that blind search in combination with deeper
sequences of lemmas quickly becomes inefficient, but at least offers
means of having to go through this computation in bulk only once,
producing in its wake leaner certificates that will run faster.

I still think that given what is shown here (with examples included),
it is not unlikely that finer elaborations can be produced on the
fly. At any rate, given the indexing structure, one can readily
prepare a "better" certificate, and in this sense the distinction is
but of minor concern.

Bounds of computation (i.e., unfoldings) would be an important
addition.