elaboration: modular structure for pair# clerks
Elaboration pairs in clerks and experts should be oblivious to the
concrete structure of the pairs, and call themselves recursively to
process both halves, thus improving modularity. For example, given
the transformation
  (pair# In1 In2) -> (pair# Out1 Out2)
the clerk/expert would itself have definitions dealing recursively
with the pair of transformations
  In1 -> Out1
  In2 -> Out2

The approach should be extended to all clerks and experts, in
particular those spelling out pair# decompositions and recompositions
such as those involving induction. Indexing structures should also be
provided for pairs in an orthogonal way, as well.

For induction and quantification, vacuous abstractions should be
extended to cover the general case, exotic as it may be.