elaboration: proper pair# certificate abstractions
Quantifiers and induction clerks involving elaboration pairs are
given the general treatment, where a vacuous abstraction in sub-
certificates is no longer assumed, and the abstraction of the pair
is extended to the application of the abstracted sub-certificates
(these are still vacuous, though).

It seems odd to me that both allClerk and someClerk, as opposed to
someExpert, produce abstractions, and this may even hint at a bug
in the kernel itself!