elaboration: local index pairs
Just as pairs in a certificate are processed separately to produce
each half of the resulting pair of continuation certificates,
clerks and experts involving indexing can organize themselves to
proceed independently, thus producing a pair of indexes that can
reach consensus ON LOCAL FORMULAS. This is fully orthogonal to the
treatment accorded to pairs heretofore.

The (at least current) exceptions are global lemmas, which need not
know anything about this scheme, and are thus directly required to
agree on the lemma to be chosen.