kernel: full index abstraction
The last remnants of the original dependencies on index definitions
in the kernel are extirpated.

Originally, the administrative FPC defines indexes as pairs of
numerical indexes (for local decisions) and boolean indexes (for
naming structures and global decisions). The leaves of those
naming structures are names themselves, and these were used to
make reference to lemmas: that is, lemma names were a very specific
subset of naming structures, themselves one of the components of
a unified index type. This dependency was injected to some limited
extent into the kernel, which made an assumption on the shape of
"lemma indexes" and performed some boilerplate "type casting."

With the simple FPC, things were set to be refactored and much
simpler, but this dependency offered some minor resistance that was
partially, but not fully, pulled out.

Kernel changes: minor, but necessary to remove the vestiges of old
dependencies.

 - Lemmas are indexed by the general 'idx' type in a fully abstract
   way, and no longer in the "boolean index" that originates in the
   administrative FPC.

 - The decideLClerk' reflects this abstraction. This is the only
   holdout of the old translation convetions.

Both FPC templates receive the corresponding adjustments to mirror
the change in protocol.

 - The simple FPC no longer needs to declare a 'boolidx' to keep
   the perverted kernel happy.

The admin FPC was the source of pollution and must be cleansed a
bit more intently.

 - First, the single idx constructor (consisting of the pair of
   numeric and boolean indexes, as aforementioned) becomes two
   constructors: one for the old pair ('idx2') and an external
   constructor that receives the old name 'idx' and takes a
   string serving as name. The name and signature of the latter
   coincide with those used in the simple FPC, which will allow
   reuse of most example files regardless of the certificate.
   (Bedwyr's module system is terrible, and this is the only way I
   see to avoid duplication without much work at the moment.)

 - The clerks and experts involving indexes now work with the index
   constructor pertaining to the type decision they are involved
   with: decideLClerk' with global decisions ('idx'), the rest
   (freezeLClerk, initRExpert, storeLClerk, decideLClerk) with
   local decisions ('idx2').

Examples are updated accordingly, i.e., lists of lemmas have their
names stated not with 'name', but with 'idx'.