The great extension renaming, redux
Following Dale's recommendation, all Bedwyr (and Abella) files are
given the *.thm extension. This makes a bit less clear what is
meant to serve as signature and what as implementation file; the
former group is given a *-sig.thm name suffix.

As always, interface changes just as uninteresting as they are
inconvenient and manifold.