kernel: add certificate unmarshalling interface
To facilitate the writing and simple use of certificates by people,
and mediate between the possibly more verbose certificates that are
useful to the machine, the kernel defines an additional
unmarshalling predicate that is called by the main proving calls to
perform the translation, if necessary, in a centralized and
convenient place. FPC templates must define this predicate
according to the needs of their certificate constructors, usually
by going from an external to an internal format.