abella: simple 'times' session
A simple session for the multiplication operation is added for
Abella. Here there are contained examples of elaborated predicates
with decision trees, as well as certificate pairs that will yield
an elaboration in the form of (for now, funnily formatted)
decision trees in the second half of the certificate pair.

Lemmas and accessory results are skipped for compactness.