Comment out debugging information
Produce a lean version by commenting out all traces. As Bedwyr
makes debugging inconvenient, I do not want to get rid of them
altogether. Instead, all the code is formatted in such a way that
leaving out the lines will not affect syntactic correctness by a
suitable trailing or adjoining of conjunction operators.

All such lines (as affect the execution of certificates) are
currently commented and denoted with the end-of-line comment
%DEBUG. In this way they are easy to toggle automatically and
without messing with Bedwyr, something I would rather not do at the
moment.