tutorial: outline and example with lemmas
The document is structured with sub-sections where appropriate.
Some missing references and links are added.

More interestingly, a more substantial example is introduced.
It makes use of lemmas and treats at some length both phases of
bipoles. It also illustrates some technical matters that will
inevitably show up when translating proofs from Abella, namely
those issues derived from references to an explicit presence of the
induction hypothesis in the list of assumptions, which is not part
of the muLJF sequent calculus. It can be easily emulated by
duplication of this assumption, even though when coupled with
unfolding its automatic translation is slightly more delicate: we
can introduce the duplication directly during translation, but it
will only be used (presumably) in some proofs by induction, and in
this case an additional unfolding will have to be inserted where
appropriate, also transparently. The process is more involved than
what we have done before and might have some strange corners. It is
luckily also easy to detect and treat by hand, and we do so in the
example.