Kernel documentation
The main kernel interfaces are documented. No profusion of
technical details, but I think all the important points are there.
Some examples may illustrate the design process, and it would be
interesting to review all documentation to see that no circular
dependencies remain and the explanations are clear in their
context.