Uncommented debugging traces in admin FPC
Missing traces in coinduction are now commented and adhere to the
conventions of debugging code. These traces are at the top of the
coinduction predicates and not very useful, except for the fact
that coinduction is a bit delicate, especially regardling naming
structures, and at some point they will probably be useful to have
while the code is being refined.