Toggle debug mode script
A Perl script to toggle on and off debugging, defined as those
lines of code in .thm files annotated with %DEBUG at the end. It
will treat all lines independently so it is sensitive to partial
states.

It wouldn't hurt to be conservative with the file names given, or
with the execution environment. We're rewriting files, reversibly,
but nonetheless.