Add ACheck script
A simple script that automates given some mild assumptions the
basic workflow of ACheck. It takes an Abella session and full paths
to both Abella (FPC flavor) and Bedwyr, and proceeds:

 - Removing any previous translation files (does not rely on the
   translator doing this for now, and assumes that the script
   itself, as well as those files, are located on the project
   root).

 - Translates the Abella session and adjusts the generated paths
   given the placement assumption above.

 - Loads and executes the checker on the shipped certificates.

It is certainly possible to manually iterate the last step, adding,
changing and removing proof assertions, and re-running as
described in the tutorial.