simple-fpc: explicit induction support
Regular induction with explicit invariants and full continuation
certificates is now supported in the simple FPC. Support extends
to blocks and chains only; pairs have not been considered in full
yet.

A test/example is added to the plus file. The objective should be
to restore all tests present in the admin FPC counterparts.

As an interesting aside, it can be seen that changes involving the
addition of new certificates are not really completely local.
Support predicates and "case analyses" must be completed, otherwise
some things may (often will) cease to work as they did, as they are
reliant on full matches in ML parlance.