fpc-simple: missing simulation examples
Everything is up to speed with the admin FPC in the main branch.
One point of interest that we see reflected in the examples is
the need to change the 'search' tactic to some kind of
'(apply 1 _ _ _ _)' due to the very constrained implementation of
'search' in the simple FPC, which we may consider upgrading.