ccs: examples for the simple FPC
Naming conventions for the admin FPC adjusted and new files for the
simple FPC created with the usual translations.

Obviously, the same theorem that currently fails to verify for the
admin FPC gets stuck with the simple FPC: it is not a matter of
certificates (even though we are going for direct equivalents), but
of the object-level logic, if I'm not mistaken.

The translations evidence one of the main differences between both
FPC templates: when we add coinduction, we need to consider that
unfolding permissions (left vs. right in admin, async vs. sync in
simple) generate slightly different certificates depending on the
kind of fixed points (least of greatest) that they involve, and
this can look a bit confusing.

Another interesting point is that for bisimulation we went into a
fair bit of detail concerning the types of releases, on the left or
right, given the Tac-like distinction and special permissions being
applied. All this is uniform in the simple FPC, which means choice
is increased, though verification time probably shouldn't in this
case. What does change is that left permissions, which were shared
for all bipoles in the admin certificate, are now given for each
bipole in the simple certificate, and this does increase (useless)
choice significantly and slows things down. But if we want to be
more precise, it suffices to move from block certificates to nested
ones and specify there which bipoles get the permissions. That is
to say, the usual tradeoff of the simple FPC, in a more homogeneous
way.