ccs: sim_mu_par works
All the examples in the Abella website now work. The proof script
made clear that there was no magic going on here, just a bit of
search (I wonder why I failed to recognize this as the problem when
first looking at it). A good example where a simple block
certificate, provided that we don't run into any cases that Bedwyr
can't handle will take care of most of our troubles.

An elaboration could even give concrete bounds so that the
generous certificate can become more of a miser, which is what all
good certificates should aspire to.