lambda-fpc: signature stubs
Generic certificate signature and stubs for implementation of a
simple FPC in lambda Prolog.

In the following we attempt a more modular definition of FPC
templates by generation from open-world style, lambda Prolog
definitions of clerks and experts, that will accummulate naturally.

Access to the logic is necessary to supply invariants, and elided
for the time being.

Also missing is a determination of the optimal structure of the
signature in particular: whether it should be split, and how.