lambda-fpc: implementation of block FPCs
A straightforward encoding of all the rules governing the use of
"block" FPCs is given in lambda Prolog. No particular attempt is
made to optimize or refactor, aiming for as direct an encoding as
possible.

In particular, we take from lambda Prolog (and the Teyjus
implementation in particular, so this may not be fully portable)
the built-in integer type and its operations; these should be
converted separately.