lambda-fpc: certificate pairs
The third set of lambda Prolog definitions completes the first
approach to modularizing the simple FPC. All the code is contained
in the three current sets of definitions (atomic, decision, pair),
if no particular attempt has been made at ultra-compact code.
Everything is explicit, which barring typos and similar errors
should be safe, and local. There are further possible refinements,
leading to increased modularity and, possibly, legibility.

There is some repetition of code and pattern matching, but being
self-contained it is tolerable for the time being. Some potential
differences with the original simple FPC are noted, which point
to design issues that need to be resolved, and refined.