fpc-lambda: guided induction
Block and continuation certificates added. Obviously, we do not
need knowledge of the logic since we are not manipulating it: we
just need to know that the type is there, and propagate
accordingly (same as with i).

The explicit syntax becomes a boon in the sense that it leaves
nothing covered: the translation of the simple certificate shows
that induction certificates are allowed to do quite a few things,
and it may be instructive to attempt to restrict these more, though
this will constrain (perhaps favorably) the style of proofs that
the certificates are intended to solve.