lambda-fpc: inductive encoding of naturals
To pave the way for as simple a translation as possible, we
currently renounce the services provided by the Teyjus standard
system, replacing all instances of integers with an inductive
encoding of the natural numbers. This simplifies clerks and
experts in some limited measure.

Naming conventions here are important, however: shall we suppose
that one such type should be generated during translation, or
should it be furnished by the kernel in which it will be embedded?
And even more, what should be the naming conventions and how should
clashes best be avoided? Obviously interface names must be
preserved, but what about the rest?