lambda-fpc: reconstruction of the simple FPC
Based on previously defined modules, the process is straightforward
and appears to work well. It remains to ensure that signatures and
modules are accumulated correctly and without duplicates, and
possibly remove redundancies in the interest of clarity. Note,
however, that while some modules are fully self-contained, others
are not, whereas yet others are pure compositions of parts, even
if there are often several ways to arrive at the final module.