lambda-fpc: Abella-approved translation example
Fixes to the atomic specification to work well with Abela and
related examples.

A word about the intended strategy (aka "plan A") is in order. In
principle, Abella has the ability to import and utilize (though the
latter is of no concern to us) specifications written in lambda
Prolog through the aptly-named Specification command at the start
of a session. Said specification is loaded into the signature and
translated into the internal term language of Abella, which its
FPC flavor already translates, at least one scenario thereof, into
Bedwyr code.

The plan then is simple: harness the built-in lambda Prolog parser
and the translation capabilities that we already have to perform
the translation in as comprehensive and headache-free a fashion as
possible.

The adjustments in this commit pertain to the seeming inability of
Abella to recognize the built-in lambda Prolog types, namely string
as needed by the indexing mechanism. To TEMPORARILY remedy this,
an alternative implementation of indexing that does away with
string constants in favor of markers and integer indexes is
substituted, and we shall live in sin until the translation by
"plan A" appears to work and can be corrected, or is discarded in
favore of an as-of-yet-undisclosed "plan B".

The alterations thus made to the atomic specification, and ported
to the "auto-generated" simple FPC and the battery of examples,
continue to work as before.