lambda-fpc: elaboration example
The elaboration FPC is extended according to the changes introduced
in atomic, on which it depends, and examples introduced in the
manual harness, but still commented due to the absence of the
suitable part of the "translation". The resulting .mod is
interesting in that it introduces slightly more complicated but
important and recurring clerk and expert patterns.

The decision specification still needs to be harmonized and its
dependencies resolved so that it can work "autonomously".