lambda-fpc: atomic example
Since currently the Abella and FPCcheck repositories are
disconnected, we create now and for temporary purposes a cached
version of the FPC template generated by the translator from the
atomic specification, with all previously mentioned problems
addressed, and examples adapted to deal with the new, adjusted
constructors. They all work as would be expected, out of the box
given the current translation.