Full implementation of pairing in lambda Prolog
Previous simplified implementations of pairing in the context of
the proof outline certificates were restricted to certain clerks
and experts relevant for the specific schemes associated with said
outline descriptions. A general implementation of pairing must
extend treatment to all clerks and experts, delegating recursively
(in the structure of the FPC) on each half of the certificate pair.