Modularization of pairing and outline elaboration
The original 'pair' FPC template implemented elaboration for the
simple proof outlines in a single package. After the generalization
of the code, the module 'pair' is left to contain exclusively this
general implementation, and the code to generate elaborations of
lemma trees on outlines is moved to its own 'lemma' module. This
affects the composition of the 'simple' FPC in its lambda Prolog
version and its Abella translation in what can essentially be seen
as a parsing-dependent reordering which given the coding
conventions followed for clerks and experts should have no
practical effect on proof search.

Other certificates derived from lambda Prolog code are not affected
by the changes, but they will benefit from the added modularity. Of
course, the quick compilations obtained by simply accumulating
'atomic' to some exceptions would not yield the same results
insofar as the resulting modules differ in their final contents.